Fixed some compiler warnings for 64bit architecture
git-svn-id: https://svn.apache.org/repos/asf/chemistry/objectivecmis/trunk@1590631 13f79535-47bb-0310-9956-ffa450edef68

+++ ObjectiveCMIS/Bindings/AtomPub/AtomPubParser/CMISAtomEntryWriter.m
@@ -202,7 +202,7 @@
             }
             default:
             {
-                CMISLogDebug(@"Property type did not match: %ld", propertyData.type);
+                CMISLogDebug(@"Property type did not match: %d", (int)propertyData.type);
                 break;
             }
         }
